I'M BACK! If your reading this, then guess what? My desktop is finally back up and running, so I will be progressively making more progress on the game, eventually getting the whole thing done in the comming month, or so. matters on how much else we decide on moving around.  we are currently still cleaning out the old appartment and so progress will still be slow, but it will be happening!
on the more important note, this update includes scripts for the menus, work done on the buttons for menus (specifically, menu buttons navigate to other menus, and back button history implemented as well)  so now the main menu will be displayed first, and player can navigate into gameplay.  still working on pause menu and shop menu.  shop menu will be the target for now, which should lead into finishing the statistics which will then connect the score and points so that we see all that running, as well as bullet collection and useage by the guns in gameplay.